Don't double-report fatal exceptions
Raven already knows how to capture and report exceptions that bring down
the application. By logging it at level ERROR, it gets double-reported.

The purpose of logging it at all is just to get it into the main log
file. (It otherwise just ends up in the journal.) So just lower the log
level.